refactor(proxy): the wire shape of a refusal gets its own module
`main.rs` crossed its line ratchet: 4976 against a ceiling of 4975. The
ratchet is doing exactly what it exists for, so this pays it rather than
raising it.

`ErrorBody`, `ApiError` and its `IntoResponse` move out verbatim — 205
lines, no behaviour change: the same variants, the same status codes, the
same stable `kind` strings. main.rs drops to 4774.

They belong together anyway. The `kind` strings ARE the client contract —
`nucleus-sdk`'s `from_error_payload` matches on them — while the message
beside them is prose that may be reworded. A file that holds both, and a
header that says which is which, makes that distinction visible to the
next person editing it; buried at line 1180 of a five-thousand-line
main.rs, it was not.
